--- packages/frontend/src/components/MarkdownRenderer/Components/NewsDrawerComponent.ts
@@ -1,15 +1,0 @@
-import MarkdownComponent from '.'
-
-export const NEWS_DRAWER_COMPONENT = MarkdownComponent({
-  styles: {
-    text: {
-      color: 'secondary.500',
-      textStyle: 'body-1',
-      fontSize: '1rem',
-    },
-    list: {
-      color: 'secondary.500',
-      marginInlineStart: '1.25em',
-    },
-  },
-})
