+++ packages/frontend/src/components/NewsDrawer/index.tsx
@@ -1,4 +1,6 @@
+import { useCallback, useState } from 'react'
 import {
+  Box,
   Button,
   Drawer,
   DrawerBody,
@@ -14,12 +16,21 @@ import {
 import NewsItem from './NewsItem'
 import { TEST_ITEM_LIST } from './TestItemList'
 
-interface NewsDrawerProps {
-  handleOpen: () => void
-}
+// this fetches the latest time from the news
+const latestNewsTimestamp =
+  TEST_ITEM_LIST.length > 0 ? TEST_ITEM_LIST[0].date.getTime().toString() : ''
 
-export default function NewsDrawer(props: NewsDrawerProps) {
-  const { handleOpen } = props
+export default function NewsDrawer() {
+  // check whether user has read and closed the news drawer
+  const [localLatestTimestamp, setLocalLatestTimestamp] = useState(
+    localStorage.getItem('news-tab-latest-timestamp'),
+  )
+
+  const handleOpen = useCallback(() => {
+    // only way to update this is to change the news or clear the local storage
+    localStorage.setItem('news-tab-latest-timestamp', latestNewsTimestamp)
+    setLocalLatestTimestamp(latestNewsTimestamp)
+  }, [latestNewsTimestamp])
 
   const { isOpen, onOpen, onClose } = useDisclosure({ onOpen: handleOpen })
 
@@ -28,6 +39,17 @@ export default function NewsDrawer(props: NewsDrawerProps) {
       <Button size="xs" variant="link" mr={4} onClick={onOpen}>
         What's new
       </Button>
+      {localLatestTimestamp !== latestNewsTimestamp && (
+        <Box
+          borderRadius="50%"
+          bg="#C05050"
+          w={1.5}
+          h={1.5}
+          ml={-5}
+          mb={3}
+          mr={2}
+        />
+      )}
       <Drawer isOpen={isOpen} placement="right" onClose={onClose} size="lg">
         <DrawerOverlay />
         <DrawerContent>
